Beef up dmmf crate to house the entire AST / serialization code.
