+++ query-engine/core/src/query_graph_builder/write/nested/mod.rs
@@ -8,10 +8,9 @@ mod update_nested;
 mod upsert_nested;
 
 use super::*;
-use schema::ConnectorContext;
-use schema_builder::constants::operations;
 use crate::{
-    query_graph::{NodeRef, QueryGraph}, ParsedInputMap,
+    query_graph::{NodeRef, QueryGraph},
+    ParsedInputMap,
 };
 use connect_nested::*;
 use connect_or_create_nested::*;
@@ -19,6 +18,8 @@ use create_nested::*;
 use delete_nested::*;
 use disconnect_nested::*;
 use prisma_models::RelationFieldRef;
+use schema::ConnectorContext;
+use schema_builder::constants::operations;
 use set_nested::*;
 use update_nested::*;
 use upsert_nested::*;
