+++ query-engine/dmmf/Cargo.toml
@@ -8,6 +8,10 @@ bigdecimal = "0.2.0"
 datamodel = { path = "../../libs/datamodel/core" }
 serde = { version = "1.0.136", features = ["derive"] }
 serde_json = "1.0.78"
+schema = { path = "../schema" }
+schema-builder = { path = "../schema-builder" }
+indexmap = { version = "1.7", features = ["serde-1"] }
+prisma-models = { path = "../prisma-models" }
 
 [dev-dependencies]
 expect-test = "1.2.2"
