+++ query-engine/dmmf/src/ast_builders/mod.rs
@@ -1,0 +1,5 @@
+mod datamodel_ast_builder;
+mod schema_ast_builder;
+
+pub use datamodel_ast_builder::*;
+pub use schema_ast_builder::*;
