+++ query-engine/dmmf/src/serialization_ast/schema_ast.rs
@@ -1,5 +1,5 @@
 use indexmap::IndexMap;
-use query_core::schema::Deprecation;
+use schema::Deprecation;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Default)]
