--- query-engine/dmmf/tests/dmmf_tests.rs
@@ -1,2 +1,0 @@
-mod common;
-mod render_to_dmmf;
