--- query-engine/dmmf/tests/render_to_dmmf/mod.rs
@@ -1,3 +1,0 @@
-pub mod parser_renderer_ast;
-pub mod parser_renderer_dml;
-pub mod render_to_dmmf;
