+++ query-engine/query-engine-node-api/src/engine.rs
@@ -2,7 +2,11 @@ use crate::{error::ApiError, logger::Logger};
 use datamodel::{dml::Datamodel, ValidatedConfiguration};
 use opentelemetry::global;
 use prisma_models::InternalDataModelBuilder;
-use query_core::{executor, schema_builder, QueryExecutor, schema::{QuerySchema, QuerySchemaRenderer}, TxId};
+use query_core::{
+    executor,
+    schema::{QuerySchema, QuerySchemaRenderer},
+    schema_builder, QueryExecutor, TxId,
+};
 use request_handlers::{GraphQLSchemaRenderer, GraphQlHandler, TxInput};
 use serde::{Deserialize, Serialize};
 use serde_json::json;
