+++ query-engine/request-handlers/src/tests/dmmf/mod.rs
@@ -1,4 +1,6 @@
-mod aggregation;
-mod helpers;
-mod ignored;
-mod unsupported;
+// Todo: Move tests to appropriate place
+
+// mod aggregation;
+// mod helpers;
+// mod ignored;
+// mod unsupported;
